test(feature): share the mock HTTP server helper to cut duplication (sonar)
The fixed-body mock-server boilerplate was duplicated between test_oapif.py and
test_wfs.py (SonarCloud new_duplicated_lines_density gate). Hoist it to a single
make_fixed_body_server in tests/feature/conftest.py and reduce both _make_server
copies to thin wrappers; also switch a dummy non-fetched href to https (S5332).
