+++ tests/feature/conftest.py
@@ -1,3 +1,7 @@
+import http.server
+import socketserver
+import threading
+from collections import Counter
 from pathlib import Path
 from typing import List, Tuple
 
@@ -7,6 +11,33 @@ from geopandas.geodataframe import GeoDataFrame
 from osgeo import ogr
 from osgeo.gdal import Dataset
 from shapely import wkt
+
+
+def make_fixed_body_server(body: str, content_type: str, path: str = ""):
+    """Start a local HTTP server returning `body` for every GET.
+
+    Shared by the WFS and OGC API – Features tests to avoid duplicating the
+    server-spin-up boilerplate. Returns ``(url, counter, httpd)``; the caller is
+    responsible for ``httpd.shutdown()`` / ``httpd.server_close()``.
+    """
+    counter: Counter[str] = Counter()
+    payload = body.encode("utf-8")
+
+    class Handler(http.server.BaseHTTPRequestHandler):
+        def do_GET(self):  # noqa: N802
+            counter["GET"] += 1
+            self.send_response(200)
+            self.send_header("Content-Type", content_type)
+            self.send_header("Content-Length", str(len(payload)))
+            self.end_headers()
+            self.wfile.write(payload)
+
+        def log_message(self, *args, **kwargs):  # noqa: N802
+            return
+
+    httpd = socketserver.ThreadingTCPServer(("127.0.0.1", 0), Handler)
+    threading.Thread(target=httpd.serve_forever, daemon=True).start()
+    return f"http://127.0.0.1:{httpd.server_address[1]}{path}", counter, httpd
 
 
 @pytest.fixture(scope="module")
