+++ tests/feature/test_oapif.py
@@ -27,10 +27,11 @@ from shapely.geometry import Point
 from pyramids.feature import FeatureCollection
 from pyramids.feature import _oapif
 from pyramids.errors import OGCAPIError
+from tests.feature.conftest import make_fixed_body_server
 
 COLLECTIONS_DOC = json.dumps(
     {
-        "links": [{"rel": "self", "href": "http://x/collections"}],
+        "links": [{"rel": "self", "href": "https://x/collections"}],
         "collections": [
             {"id": "lakes", "title": "Lakes"},
             {"id": "roads", "title": "Roads"},
@@ -50,26 +51,8 @@ def _clear_collections_cache():
 
 
 def _make_server(body: str, content_type: str = "application/json"):
-    """Start a local HTTP server returning `body` for every GET; yield (url, counter, httpd)."""
-    counter: Counter[str] = Counter()
-    payload = body.encode("utf-8")
-
-    class Handler(http.server.BaseHTTPRequestHandler):
-        def do_GET(self):  # noqa: N802
-            counter["GET"] += 1
-            self.send_response(200)
-            self.send_header("Content-Type", content_type)
-            self.send_header("Content-Length", str(len(payload)))
-            self.end_headers()
-            self.wfile.write(payload)
-
-        def log_message(self, *args, **kwargs):  # noqa: N802
-            return
-
-    httpd = socketserver.ThreadingTCPServer(("127.0.0.1", 0), Handler)
-    port = httpd.server_address[1]
-    threading.Thread(target=httpd.serve_forever, daemon=True).start()
-    return f"http://127.0.0.1:{port}", counter, httpd
+    """Local HTTP server returning `body` for every GET; returns (url, counter, httpd)."""
+    return make_fixed_body_server(body, content_type)
 
 
 @pytest.fixture
